# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend Copy",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://65.2.142.187:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/role/create": {
			"post": {
				"tags": ["Admin Role"],
				"summary": "CreateRole",
				"description": "Create Role",
				"operationId": "CreateRole",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleRequest"
							},
							"example": {
								"name": "EDITOR",
								"permissions": ["6110d5db84d4803c28737723"]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/role/get": {
			"get": {
				"tags": ["Admin Role"],
				"summary": "Getrole",
				"description": "Get role",
				"operationId": "Getrole",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/role/update/{id}": {
			"put": {
				"tags": ["Admin Role"],
				"summary": "updateRole",
				"description": "update Role",
				"operationId": "updateRole",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateRoleRequest"
							},
							"example": {
								"name": "SUPER_USER",
								"createdBy": "6107cc835e94d0bc7dd9f1c2",
								"permissions": [
									"610d0b608308ca8cce20c41f",
									"610d0b608308ca8cce20c420",
									"610d0b608308ca8cce20c421",
									"610d0b608308ca8cce20c422",
									"610d0b608308ca8cce20c423",
									"610d0b608308ca8cce20c424",
									"610d0b608308ca8cce20c425",
									"610d0b608308ca8cce20c426",
									"610d0b608308ca8cce20c427",
									"610d0b608308ca8cce20c428",
									"610d0b608308ca8cce20c429",
									"610d0b608308ca8cce20c42a",
									"610d0b608308ca8cce20c42b",
									"610d0b608308ca8cce20c42c",
									"610d0b608308ca8cce20c42d",
									"610d0b608308ca8cce20c42e",
									"610d0b608308ca8cce20c42f",
									"610d0b608308ca8cce20c430",
									"610d0b608308ca8cce20c431",
									"610d0b608308ca8cce20c432",
									"610d0b608308ca8cce20c433",
									"610d0b608308ca8cce20c434",
									"610d0b608308ca8cce20c435",
									"610d0b608308ca8cce20c436",
									"610d0b608308ca8cce20c437",
									"610d0b608308ca8cce20c438",
									"610d0b608308ca8cce20c439",
									"610d0b608308ca8cce20c43a",
									"610d0b608308ca8cce20c43b",
									"610d0b608308ca8cce20c43c",
									"610d0b608308ca8cce20c43d",
									"610d0b608308ca8cce20c43e",
									"610d0b608308ca8cce20c43f",
									"610d0b608308ca8cce20c440",
									"610d0b608308ca8cce20c441",
									"610d0b608308ca8cce20c442",
									"610d0b608308ca8cce20c443",
									"610d0b608308ca8cce20c444",
									"610d0b608308ca8cce20c445",
									"610d0b608308ca8cce20c446",
									"610d0b608308ca8cce20c447",
									"610d0b608308ca8cce20c448",
									"610d0b618308ca8cce20c449",
									"610d0b618308ca8cce20c44a",
									"610d0b618308ca8cce20c44b",
									"610d0b618308ca8cce20c44c",
									"610d0b618308ca8cce20c44d",
									"610d0b618308ca8cce20c44e",
									"610d0b618308ca8cce20c44f",
									"610d0b618308ca8cce20c450",
									"610d0b618308ca8cce20c451",
									"610d0b618308ca8cce20c452",
									"610d0b618308ca8cce20c453",
									"610d0b618308ca8cce20c454",
									"610d0b618308ca8cce20c455",
									"6110c1ed170cf61a4c8d3e47",
									"6110d11772f7ae3890564085",
									"6110d0b3449850384837abc4"
								]
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/role/delete/{id}": {
			"delete": {
				"tags": ["Admin Role"],
				"summary": "DeleteRole",
				"description": "Delete Role",
				"operationId": "DeleteRole",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/admin-user/status/{id}": {
			"put": {
				"tags": ["Admin User"],
				"summary": "UpdateAdminUser",
				"description": "Update Admin User",
				"operationId": "UpdateAdminUser",
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/admin-user/forget-password": {
			"post": {
				"tags": ["Admin User"],
				"summary": "ForgetPassword",
				"description": "Forget Password",
				"operationId": "ForgetPassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgetPasswordRequest"
							},
							"example": {
								"email": "tushar18@mailinator.com"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false
			}
		},
		"/admin-user/reset-password": {
			"put": {
				"tags": ["Admin User"],
				"summary": "ResetPassword",
				"description": "Reset Password",
				"operationId": "ResetPassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							},
							"example": {
								"email": "tushar18@mailinator.com",
								"newPassword": "123456"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateRoleRequest": {
				"title": "CreateRoleRequest",
				"required": ["name", "permissions"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"permissions": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"name": "EDITOR",
					"permissions": ["6110d5db84d4803c28737723"]
				}
			},
			"updateRoleRequest": {
				"title": "updateRoleRequest",
				"required": ["name", "createdBy", "permissions"],
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"permissions": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"name": "SUPER_USER",
					"createdBy": "6107cc835e94d0bc7dd9f1c2",
					"permissions": [
						"610d0b608308ca8cce20c41f",
						"610d0b608308ca8cce20c420",
						"610d0b608308ca8cce20c421",
						"610d0b608308ca8cce20c422",
						"610d0b608308ca8cce20c423",
						"610d0b608308ca8cce20c424",
						"610d0b608308ca8cce20c425",
						"610d0b608308ca8cce20c426",
						"610d0b608308ca8cce20c427",
						"610d0b608308ca8cce20c428",
						"610d0b608308ca8cce20c429",
						"610d0b608308ca8cce20c42a",
						"610d0b608308ca8cce20c42b",
						"610d0b608308ca8cce20c42c",
						"610d0b608308ca8cce20c42d",
						"610d0b608308ca8cce20c42e",
						"610d0b608308ca8cce20c42f",
						"610d0b608308ca8cce20c430",
						"610d0b608308ca8cce20c431",
						"610d0b608308ca8cce20c432",
						"610d0b608308ca8cce20c433",
						"610d0b608308ca8cce20c434",
						"610d0b608308ca8cce20c435",
						"610d0b608308ca8cce20c436",
						"610d0b608308ca8cce20c437",
						"610d0b608308ca8cce20c438",
						"610d0b608308ca8cce20c439",
						"610d0b608308ca8cce20c43a",
						"610d0b608308ca8cce20c43b",
						"610d0b608308ca8cce20c43c",
						"610d0b608308ca8cce20c43d",
						"610d0b608308ca8cce20c43e",
						"610d0b608308ca8cce20c43f",
						"610d0b608308ca8cce20c440",
						"610d0b608308ca8cce20c441",
						"610d0b608308ca8cce20c442",
						"610d0b608308ca8cce20c443",
						"610d0b608308ca8cce20c444",
						"610d0b608308ca8cce20c445",
						"610d0b608308ca8cce20c446",
						"610d0b608308ca8cce20c447",
						"610d0b608308ca8cce20c448",
						"610d0b618308ca8cce20c449",
						"610d0b618308ca8cce20c44a",
						"610d0b618308ca8cce20c44b",
						"610d0b618308ca8cce20c44c",
						"610d0b618308ca8cce20c44d",
						"610d0b618308ca8cce20c44e",
						"610d0b618308ca8cce20c44f",
						"610d0b618308ca8cce20c450",
						"610d0b618308ca8cce20c451",
						"610d0b618308ca8cce20c452",
						"610d0b618308ca8cce20c453",
						"610d0b618308ca8cce20c454",
						"610d0b618308ca8cce20c455",
						"6110c1ed170cf61a4c8d3e47",
						"6110d11772f7ae3890564085",
						"6110d0b3449850384837abc4"
					]
				}
			},
...
			"ForgetPasswordRequest": {
				"title": "ForgetPasswordRequest",
				"required": ["email"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"example": {
					"email": "tushar18@mailinator.com"
				}
			},
			"ResetPasswordRequest": {
				"title": "ResetPasswordRequest",
				"required": ["email", "newPassword"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"example": {
					"email": "tushar18@mailinator.com",
					"newPassword": "123456"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}